✅ Create separate tokenLogger action
