+++ src/store/ConfigureStore.jsx
@@ -1,4 +1,5 @@
 import { configureStore } from '@reduxjs/toolkit';
+import tokenLogger from '../data-api/tokenLogger';
 import userSlice from '../redux/UserSlice';
 
 const rootReducer = {
@@ -6,5 +7,6 @@ const rootReducer = {
 };
 
 const store = configureStore({ reducer: rootReducer });
+store.dispatch(tokenLogger());
 
 export default store;
